추가 : admin 외 수정사항

+++ src/users/admin/admin.controller.ts
@@ -4,26 +4,56 @@ import {
   UseGuards,
   Delete,
   Param,
+  Req,
   ForbiddenException,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { Roles } from './roles.decorator';
 import { RolesGuard } from './roles.guard';
 import { UsersService } from '../users.service';
 
-@Controller('admin')
-@UseGuards(RolesGuard)
-export class UsersController {
+@Controller('admin') // @UseGuards(RolesGuard)
+export class AdminController {
   constructor(private readonly usersService: UsersService) {}
 
   @Get('user-list')
-  @Roles(['admin'])
-  async getUserList() {
+  async getUserList(@Req() request) {
+    const accessToken = request.headers.authorization;
+
+    if (!accessToken) {
+      throw new UnauthorizedException('토큰이 없습니다');
+    }
+
+    const userId = await this.usersService.verifyAccessTokenAndGetUserId(
+      accessToken,
+    );
+    console.log(userId);
+
+    const user = await this.usersService.getUser(userId);
+
+    if (user.roles !== 'admin') {
+      throw new ForbiddenException('admin 계정이 아닙니다.');
+    }
+
     return this.usersService.getUserList();
   }
 
   @Delete('users/:user_id')
-  @Roles(['admin'])
-  async deleteUser(@Param('user_id') user_id: string) {
+  async deleteUser(@Param('user_id') user_id, @Req() request) {
+    const accessToken = request.headers.authorization?.split(' ')[1];
+
+    if (!accessToken) {
+      throw new UnauthorizedException('토큰이 없습니다');
+    }
+
+    const userId = await this.usersService.verifyAccessTokenAndGetUserId(
+      accessToken,
+    );
+    const user = await this.usersService.getUser(userId);
+
+    if (user.roles !== 'admin') {
+      throw new ForbiddenException('admin 계정이 아닙니다.');
+    }
     return this.usersService.deleteUser(user_id);
   }
 }
